rename the package to kube
