+++ infrastructure/cli/pkg/kube/BUILD
@@ -10,3 +10,14 @@ go_library(
         "@io_k8s_client_go//tools/clientcmd:go_default_library",
     ],
 )
+
+go_library(
+    name = "kube",
+    srcs = ["context.go"],
+    importpath = "cli/pkg/kube",
+    visibility = ["//visibility:public"],
+    deps = [
+        "@io_k8s_client_go//kubernetes:go_default_library",
+        "@io_k8s_client_go//tools/clientcmd:go_default_library",
+    ],
+)
