+++ infrastructure/cli/pkg/providers/aws/aws.go
@@ -1,7 +1,7 @@
 package aws
 
 import (
-	"cli/pkg/k8s_cluster"
+	"cli/pkg/kube"
 	"fmt"
 	"os"
 )
@@ -9,10 +9,10 @@ import (
 type Aws struct {
 }
 
-func (a *Aws) Provision() (k8s_cluster.Context, error) {
+func (a *Aws) Provision() (kube.KubeCtx, error) {
 	// Use this to
 	//clientcmd.NewNonInteractiveClientConfig()
 	fmt.Println("minikube provider not yet implemented")
 	os.Exit(1)
-	return k8s_cluster.Context{}, nil
+	return kube.KubeCtx{}, nil
 }
